# Adds profile guided optimization script.