--- a/.gitignore
+++ b/.gitignore
@@ -227,3 +227,5 @@ cython_debug/
 !crates/ruff_python_resolver/resources/test/airflow/venv/lib
 !crates/ruff_python_resolver/resources/test/airflow/venv/lib/python3.11/site-packages/_watchdog_fsevents.cpython-311-darwin.so
 !crates/ruff_python_resolver/resources/test/airflow/venv/lib/python3.11/site-packages/orjson/orjson.cpython-311-darwin.so
+
+clones