<!--
Thank you for contributing to Ruff! To help us out with reviewing, please consider the following:

- Does this pull request include a summary of the change? (See below.)
- Does this pull request include a descriptive title?
- Does this pull request include references to any relevant issues?
-->

## Summary

This PR adds a script to generate a pgo optimized binary of ruff, see #7055 for more detail.
I initially only implemented the script on `x86_64-unknown-linux-gnu` so people can try it out, if we think this is more interesting it should probably be added to the release pipeline. from what I can see you get a general 3-10% improvement of runtime speed then running the pgo-binaries, as well as a somewhat smaller binary size. 

The script fetches a large popular python codebases and uses that as training data (see `pgo_profile.json`),

this is an example of the performance for a cold run on pytorch's source, where the pgo is 12% faster.
```
✦ ❯ hyperfine "./target/x86_64-unknown-linux-gnu/release/ruff check -e -n  clones/pytorch/" "./target/release/ruff check -e -n  clones/pytorch/" --warmup 100 -r 1000
Benchmark 1: ./target/x86_64-unknown-linux-gnu/release/ruff check -e -n  clones/pytorch/
  Time (mean ± σ):     169.4 ms ±   6.4 ms    [User: 2942.4 ms, System: 244.4 ms]
  Range (min … max):   151.0 ms … 212.7 ms    1000 runs

Benchmark 2: ./target/release/ruff check -e -n  clones/pytorch/
  Time (mean ± σ):     190.5 ms ±   7.5 ms    [User: 3641.0 ms, System: 239.7 ms]
  Range (min … max):   168.1 ms … 235.4 ms    1000 runs

Summary
  ./target/x86_64-unknown-linux-gnu/release/ruff check -e -n  clones/pytorch/ ran
    1.12 ± 0.06 times faster than ./target/release/ruff check -e -n  clones/pytorch/
```

## Test Plan

<!-- How was it tested? -->