@MichaReiser 2025-04-25
Sorry, I haven't had time to look at this yet (I was out all last week and we're getting close to an important release. That's why my time is a bit limited at the moment). But this PR is still on my todo list